app/requirement.txt
fastapi==0.103.1
uvicorn==0.22.0
scikit-learn==1.2.2
numpy==1.23.5
joblib==1.2.0
pydantic==2.0.3
slowapi==0.1.5 
requests==2.31.0 
streamlit
pandasai==2.2.15
